Moved code from single express.js file to a more expandable system
