--- app/express/passport/config/keys.js
@@ -1,2 +1,0 @@
-// Contains all API keys for passport
-// THIS SHOULD NOT GET PUSHED ONTO GITHUB
