--- app/express/passport/config/passport-setup.js
@@ -1,3 +1,0 @@
-// The place where you setup strategies.
-// require("./keys");
-// Require this in the main file(express.js) so that the strategies get setup.
